更新推广bug 、货架和仓库更新数据协议

+++ city/gs/src/main/java/Game/Prob.java
@@ -3,7 +3,7 @@ package Game;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class Prob {
-    public static boolean success(int w, int radix) {
+    public static boolean success(double w, int radix) {
         return Math.random()*radix < w;
     }
     // [l, r)
